requirements.txt
# requirements.txt (main)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
psutil

# AI/ML dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
diffusers>=0.24.0
pillow>=10.1.0
numpy>=1.24.0

# Additional utilities
requests>=2.31.0
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
ruff>=0.1.6
black>=23.11.0

# Optional for future phases
gradio>=4.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# backend/requirements.txt (specific backend deps)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart


